pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "c4b-nbs-pipeline"
version = "0.1.0"
description = "Open-source data ingestion pipeline for Copernicus Sentinel-1/2 imagery integration with IoT environmental sensor streams."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Maria Emilia Mazzone", email = "mazzone@cloud4business.it"},
    {name = "Stefano Falcone", email = "falcone@cloud4business.it"},
]
keywords = [
    "nature-based-solutions",
    "copernicus",
    "sentinel",
    "earth-observation",
    "iot",
    "lorawan",
    "mqtt",
    "environmental-monitoring",
    "data-ingestion",
    "edge-computing",
    "digital-sovereignty",
    "FAIR-data",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
    "Topic :: Scientific/Engineering :: Hydrology",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "geopandas>=0.14",
    "xarray>=2023.10",
    "rioxarray>=0.15",
    "rasterio>=1.3",
    "shapely>=2.0",
    "pyproj>=3.6",
    "pystac-client>=0.7",
    "paho-mqtt>=1.6",
    "requests>=2.31",
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "scipy>=1.11",
    "zarr>=2.16",
    "fsspec>=2023.10",
    "tqdm>=4.66",
    "structlog>=23.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.6",
]
lorawan = [
    "chirpstack-api>=4.0",
    "grpcio>=1.60",
]
docs = [
    "sphinx>=7.2",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
]

[project.urls]
Homepage = "https://github.com/cloud4business/c4b-nbs-pipeline"
Documentation = "https://c4b-nbs-pipeline.readthedocs.io"
Repository = "https://github.com/cloud4business/c4b-nbs-pipeline"
Issues = "https://github.com/cloud4business/c4b-nbs-pipeline/issues"

[project.scripts]
c4b-pipeline = "c4b_nbs_pipeline.cli:main"

[tool.setuptools.packages.find]
include = ["c4b_nbs_pipeline*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=c4b_nbs_pipeline --cov-report=term-missing"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
